pyproject.toml
[project]
name = "coderstation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.11"
dependencies = [
    "django>=6.0",
    "django-ninja>=1.5.1",
    "django-ninja-extra>=0.30.8",
    "django-ninja-jwt>=5.4.2",
    "django-shortuuidfield>=0.1.3",
    "mysqlclient>=2.2.7",
    "ninja-schema>=0.14.3",
    "orjson>=3.11.5",
    "pillow>=12.0.0",
]
